made the player face right directions

+++ core/src/com/norwand/game/utility/Directions.java
@@ -2,9 +2,9 @@ package com.norwand.game.utility;
 
 public abstract class Directions {
 
-    public static final int NORTH = 0;
-    public static final int SOUTH = 1;
-    public static final int EAST = 2;
-    public static final int WEST = 3;
+    public static final byte NORTH = 0;
+    public static final byte SOUTH = 1;
+    public static final byte EAST = 2;
+    public static final byte WEST = 3;
 
 }
